Added tests for the claims
