+++ Claims/ClaimExtensions.cs
@@ -1,0 +1,6 @@
+﻿namespace Defender {
+	/// <summary>
+	/// Extensions to <see cref="Claim{T}"/> for further testing.
+	/// </summary>
+	public static partial class ClaimExtensions { }
+}
