+++ Claims/SequenceEquals.cs
@@ -23,14 +23,17 @@ namespace Defender {
 		/// <summary>
 		/// Does the <see cref="Claim{T}"/>'d sequence equal the <paramref name="expected"/>?
 		/// </summary>
-		/// <typeparam name="E">The sequence type.</typeparam>
-		/// <typeparam name="T">The type of the elements in the sequence.</typeparam>
+		/// <typeparam name="T">The type of the sequence.</typeparam>
 		/// <param name="claim">The <see cref="Claim{T}"/>.</param>
 		/// <param name="expected">The expected sequence.</param>
 		/// <returns>The calling <paramref name="claim"/>.</returns>
-		public static Claim<E> SequenceEquals<E, T>(this Claim<E> claim, E expected) where E : IEnumerable<T> {
-			if (!expected.SequenceEqual(claim.Value)) {
-				throw new EqualException(expected, claim.Value);
+		public static Claim<T> SequenceEquals<T>(this Claim<T> claim, T expected) where T : IEnumerable {
+			IEnumerator actEnum = claim.Value.GetEnumerator();
+			IEnumerator expEnum = expected.GetEnumerator();
+			while (actEnum.MoveNext() && expEnum.MoveNext()) {
+				if (!expEnum.Current.Equals(actEnum.Current)) {
+					throw new EqualException(expected, claim.Value);
+				}
 			}
 			return claim;
 		}
