--- Trials/ClaimExtensions.cs
@@ -1,6 +1,0 @@
-﻿namespace Defender {
-	/// <summary>
-	/// Extensions to <see cref="Claim{T}"/> for further testing.
-	/// </summary>
-	public static partial class ClaimExtensions { }
-}
