+++ Trials/Trial.cs
@@ -20,11 +20,6 @@ namespace Defender {
 		public static Claim<T[]> Claim<T>(T[] array) => new Claim<T[]>(array);
 
 		/// <summary>
-		/// Claims to be made about a sequence.
-		/// </summary>
-		public static Claim<E, T> Claim<E, T>(E sequence) where E : IEnumerable<T> => new Claim<E, T>(sequence);
-
-		/// <summary>
 		/// Claims to be made about an <see cref="Action"/>.
 		/// </summary>
 		public static IActionClaim Claim(Action action) => new ActionClaim(action);
