+++ Trials/Trials.csproj
@@ -22,7 +22,7 @@
 
 	<ItemGroup>
 	  <PackageReference Include="System.Memory" Version="4.5.3" />
-	  <PackageReference Include="xunit" Version="2.4.1" />
+	  <PackageReference Include="xunit.assert" Version="2.4.1" />
 	</ItemGroup>
 
 </Project>
