+++ docs/api/Defender.Claim-1.html
@@ -223,6 +223,7 @@
   <h5 class="decalaration">Declaration</h5>
   <div class="codewrapper">
     <pre><code class="lang-csharp hljs">public Claim&lt;T&gt; Equals&lt;Te&gt;(Te expected)
+
     where Te : IEquatable&lt;T&gt;</code></pre>
   </div>
   <h5 class="parameters">Parameters</h5>
