+++ docs/api/Defender.Claim-2.html
@@ -107,6 +107,7 @@
   <h5 id="Defender_Claim_2_syntax">Syntax</h5>
   <div class="codewrapper">
     <pre><code class="lang-csharp hljs">public sealed class Claim&lt;E, T&gt;
+
     where E : IEnumerable&lt;T&gt;</code></pre>
   </div>
   <h5 class="typeParameters">Type Parameters</h5>
@@ -228,6 +229,7 @@
   <h5 class="decalaration">Declaration</h5>
   <div class="codewrapper">
     <pre><code class="lang-csharp hljs">public Claim&lt;E, T&gt; Equals&lt;Ee&gt;(Ee expected)
+
     where Ee : IEquatable&lt;E&gt;</code></pre>
   </div>
   <h5 class="parameters">Parameters</h5>
