+++ docs/api/Defender.Guard.html
@@ -168,6 +168,7 @@
   <h5 class="decalaration">Declaration</h5>
   <div class="codewrapper">
     <pre><code class="lang-csharp hljs">public static void Empty&lt;C&gt;(C collection, string name)
+
     where C : ICollection</code></pre>
   </div>
   <h5 class="parameters">Parameters</h5>
@@ -281,6 +282,7 @@
   <h5 class="decalaration">Declaration</h5>
   <div class="codewrapper">
     <pre><code class="lang-csharp hljs">public static void Empty&lt;C, T&gt;(C collection, string name)
+
     where C : ICollection&lt;T&gt;</code></pre>
   </div>
   <h5 class="parameters">Parameters</h5>
@@ -473,6 +475,7 @@
   <h5 class="decalaration">Declaration</h5>
   <div class="codewrapper">
     <pre><code class="lang-csharp hljs">public static void Equal&lt;C&gt;(C collection, string name, int size)
+
     where C : ICollection</code></pre>
   </div>
   <h5 class="parameters">Parameters</h5>
@@ -598,6 +601,7 @@
   <h5 class="decalaration">Declaration</h5>
   <div class="codewrapper">
     <pre><code class="lang-csharp hljs">public static void Equal&lt;C, T&gt;(C collection, string name, int size)
+
     where C : ICollection&lt;T&gt;</code></pre>
   </div>
   <h5 class="parameters">Parameters</h5>
@@ -666,6 +670,7 @@
   <h5 class="decalaration">Declaration</h5>
   <div class="codewrapper">
     <pre><code class="lang-csharp hljs">public static void GreaterThan&lt;T&gt;(T value, string name, T lower)
+
     where T : IComparable&lt;T&gt;</code></pre>
   </div>
   <h5 class="parameters">Parameters</h5>
@@ -729,6 +734,7 @@
   <h5 class="decalaration">Declaration</h5>
   <div class="codewrapper">
     <pre><code class="lang-csharp hljs">public static void GreaterThan&lt;C&gt;(C collection, string name, int lower)
+
     where C : ICollection</code></pre>
   </div>
   <h5 class="parameters">Parameters</h5>
@@ -854,6 +860,7 @@
   <h5 class="decalaration">Declaration</h5>
   <div class="codewrapper">
     <pre><code class="lang-csharp hljs">public static void GreaterThan&lt;C, T&gt;(C collection, string name, int lower)
+
     where C : ICollection&lt;T&gt;</code></pre>
   </div>
   <h5 class="parameters">Parameters</h5>
@@ -922,6 +929,7 @@
   <h5 class="decalaration">Declaration</h5>
   <div class="codewrapper">
     <pre><code class="lang-csharp hljs">public static void GreaterThanOrEqualTo&lt;T&gt;(T value, string name, T lower)
+
     where T : IComparable&lt;T&gt;</code></pre>
   </div>
   <h5 class="parameters">Parameters</h5>
@@ -985,6 +993,7 @@
   <h5 class="decalaration">Declaration</h5>
   <div class="codewrapper">
     <pre><code class="lang-csharp hljs">public static void GreaterThanOrEqualTo&lt;C&gt;(C collection, string name, int lower)
+
     where C : ICollection</code></pre>
   </div>
   <h5 class="parameters">Parameters</h5>
@@ -1110,6 +1119,7 @@
   <h5 class="decalaration">Declaration</h5>
   <div class="codewrapper">
     <pre><code class="lang-csharp hljs">public static void GreaterThanOrEqualTo&lt;C, T&gt;(C collection, string name, int lower)
+
     where C : ICollection&lt;T&gt;</code></pre>
   </div>
   <h5 class="parameters">Parameters</h5>
@@ -1178,6 +1188,7 @@
   <h5 class="decalaration">Declaration</h5>
   <div class="codewrapper">
     <pre><code class="lang-csharp hljs">public static void LesserThan&lt;T&gt;(T value, string name, T upper)
+
     where T : IComparable&lt;T&gt;</code></pre>
   </div>
   <h5 class="parameters">Parameters</h5>
@@ -1241,6 +1252,7 @@
   <h5 class="decalaration">Declaration</h5>
   <div class="codewrapper">
     <pre><code class="lang-csharp hljs">public static void LesserThan&lt;C&gt;(C collection, string name, int upper)
+
     where C : ICollection</code></pre>
   </div>
   <h5 class="parameters">Parameters</h5>
@@ -1366,6 +1378,7 @@
   <h5 class="decalaration">Declaration</h5>
   <div class="codewrapper">
     <pre><code class="lang-csharp hljs">public static void LesserThan&lt;C, T&gt;(C collection, string name, int upper)
+
     where C : ICollection&lt;T&gt;</code></pre>
   </div>
   <h5 class="parameters">Parameters</h5>
@@ -1434,6 +1447,7 @@
   <h5 class="decalaration">Declaration</h5>
   <div class="codewrapper">
     <pre><code class="lang-csharp hljs">public static void LesserThanOrEqualTo&lt;T&gt;(T value, string name, T upper)
+
     where T : IComparable&lt;T&gt;</code></pre>
   </div>
   <h5 class="parameters">Parameters</h5>
@@ -1497,6 +1511,7 @@
   <h5 class="decalaration">Declaration</h5>
   <div class="codewrapper">
     <pre><code class="lang-csharp hljs">public static void LesserThanOrEqualTo&lt;C&gt;(C collection, string name, int upper)
+
     where C : ICollection</code></pre>
   </div>
   <h5 class="parameters">Parameters</h5>
@@ -1622,6 +1637,7 @@
   <h5 class="decalaration">Declaration</h5>
   <div class="codewrapper">
     <pre><code class="lang-csharp hljs">public static void LesserThanOrEqualTo&lt;C, T&gt;(C collection, string name, int upper)
+
     where C : ICollection&lt;T&gt;</code></pre>
   </div>
   <h5 class="parameters">Parameters</h5>
@@ -1730,6 +1746,7 @@
   <h5 class="decalaration">Declaration</h5>
   <div class="codewrapper">
     <pre><code class="lang-csharp hljs">public static void NotEmpty&lt;C&gt;(C collection, string name)
+
     where C : ICollection</code></pre>
   </div>
   <h5 class="parameters">Parameters</h5>
@@ -1843,6 +1860,7 @@
   <h5 class="decalaration">Declaration</h5>
   <div class="codewrapper">
     <pre><code class="lang-csharp hljs">public static void NotEmpty&lt;C, T&gt;(C collection, string name)
+
     where C : ICollection&lt;T&gt;</code></pre>
   </div>
   <h5 class="parameters">Parameters</h5>
@@ -2035,6 +2053,7 @@
   <h5 class="decalaration">Declaration</h5>
   <div class="codewrapper">
     <pre><code class="lang-csharp hljs">public static void NotEqual&lt;C&gt;(C collection, string name, int size)
+
     where C : ICollection</code></pre>
   </div>
   <h5 class="parameters">Parameters</h5>
@@ -2160,6 +2179,7 @@
   <h5 class="decalaration">Declaration</h5>
   <div class="codewrapper">
     <pre><code class="lang-csharp hljs">public static void NotEqual&lt;C, T&gt;(C collection, string name, int size)
+
     where C : ICollection&lt;T&gt;</code></pre>
   </div>
   <h5 class="parameters">Parameters</h5>
@@ -2228,6 +2248,7 @@
   <h5 class="decalaration">Declaration</h5>
   <div class="codewrapper">
     <pre><code class="lang-csharp hljs">public static void NotNull&lt;T&gt;(T value, string name)
+
     where T : class</code></pre>
   </div>
   <h5 class="parameters">Parameters</h5>
@@ -2285,6 +2306,7 @@
   <h5 class="decalaration">Declaration</h5>
   <div class="codewrapper">
     <pre><code class="lang-csharp hljs">public static void NotNull&lt;T&gt;(T? value, string name)
+
     where T : struct</code></pre>
   </div>
   <h5 class="parameters">Parameters</h5>
@@ -2398,6 +2420,7 @@
   <h5 class="decalaration">Declaration</h5>
   <div class="codewrapper">
     <pre><code class="lang-csharp hljs">public static void Valid&lt;E&gt;(E value, string name)
+
     where E : Enum</code></pre>
   </div>
   <h5 class="parameters">Parameters</h5>
@@ -3096,6 +3119,7 @@
   <h5 class="decalaration">Declaration</h5>
   <div class="codewrapper">
     <pre><code class="lang-csharp hljs">public static void Within&lt;T&gt;(T value, string name, T lower, T upper, bool inclusive)
+
     where T : IComparable&lt;T&gt;</code></pre>
   </div>
   <h5 class="parameters">Parameters</h5>
