+++ docs/api/Defender.IActionClaim.html
@@ -99,6 +99,7 @@
   <h5 class="decalaration">Declaration</h5>
   <div class="codewrapper">
     <pre><code class="lang-csharp hljs">IActionClaim Throws&lt;E&gt;()
+
     where E : Exception</code></pre>
   </div>
   <h5 class="returns">Returns</h5>
