+++ docs/api/Defender.Trial.html
@@ -387,6 +387,7 @@
   <h5 class="decalaration">Declaration</h5>
   <div class="codewrapper">
     <pre><code class="lang-csharp hljs">public static Claim&lt;E, T&gt; Claim&lt;E, T&gt;(E sequence)
+
     where E : IEnumerable&lt;T&gt;</code></pre>
   </div>
   <h5 class="parameters">Parameters</h5>
